update settings command
